requirement.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-bootstrap4==22.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dynaconf==3.1.9
filelock==3.7.1
hvac==0.11.2
idna==3.3
loguru==0.6.0
oauthlib==3.2.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.11
validators==0.20.0
virtualenv==20.16.0
win32-setctime==1.1.0
